requirements.txt
torch
torchvision
torchaudio
pytorch-lightning
sentencepiece
av
hydra-core
six
matplotlib
scikit-learn